refactored user routes using controllers
